Use PropTypes.elementType for componentTag
Previously, the type checking looked for a React element (eg an
instantiated object, `<MyComponent />`) but the implementation required
a type (eg `MyComponent`) in order to work. This updates the type
checking and the documentation and adds a test to ensure the type
is correct when a React element type is passed to `componentTag`.